requirements.txt
fake-useragent==0.1.11
factory_boy==3.2.0
ftfy==5.8
html2text==2020.1.16
jellyfish>=0.7
moment==0.12.1
pandas==1.1.5
peewee==3.12.0
scrapy==2.4.1
ipython==7.16.1
jedi==0.17.2
jupyterlab==2.1.3
matplotlib==3.3.3
python-telegram-bot==13.1
